refactor: extract navigation menu components
